Move all of ServerFile into subfolder
This makes the client-representation of the "server-side" of the house in a single folder.
Also drop the confusing "content.ts" class
