+++ src/repository/repository.ts
@@ -2,8 +2,8 @@ import Followup, { andThen } from "@util/promise/followup";
 import FollowupList from "@util/promise/followuplist";
 import Util from "@util/util";
 import XML from "@util/xml";
-import Metadata from "./metadata";
-import ServerFile from "./serverfile";
+import Metadata from "./serverfile/metadata";
+import ServerFile from "./serverfile/serverfile";
 import CaseFile from "./serverfile/casefile";
 import CFIDFile from "./serverfile/cfidfile";
 import DimensionsFile from "./serverfile/dimensionsfile";
