+++ src/repository/serverfile/cfidfile.ts
@@ -1,5 +1,5 @@
 import CaseFileDefinitionDefinition from "@repository/definition/cfid/casefileitemdefinitiondefinition";
-import ServerFile from "../serverfile";
+import ServerFile from "./serverfile";
 
 export default class CFIDFile extends ServerFile<CaseFileDefinitionDefinition> {
     createModelDefinition() {
