+++ src/repository/serverfile/dimensionsfile.ts
@@ -1,5 +1,5 @@
 import Dimensions from "@repository/definition/dimensions/dimensions";
-import ServerFile from "../serverfile";
+import ServerFile from "./serverfile";
 
 export default class DimensionsFile extends ServerFile<Dimensions> {
     createModelDefinition() {
