+++ src/repository/serverfile/humantaskfile.ts
@@ -1,5 +1,5 @@
 import HumanTaskModelDefinition from "@repository/definition/humantask/humantaskmodeldefinition";
-import ServerFile from "../serverfile";
+import ServerFile from "./serverfile";
 
 export default class HumanTaskFile extends ServerFile<HumanTaskModelDefinition> {
     createModelDefinition() {
