+++ src/repository/serverfile/processfile.ts
@@ -1,5 +1,5 @@
 import ProcessModelDefinition from "@repository/definition/process/processmodeldefinition";
-import ServerFile from "../serverfile";
+import ServerFile from "./serverfile";
 
 export default class ProcessFile extends ServerFile<ProcessModelDefinition> {
     createModelDefinition() {
